backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
solana==0.30.2
base58==2.1.1
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0 